Extract randombytes impl
