+++ Makefile
@@ -1,6 +1,6 @@
 CC = gcc
 TARGET = pam_bacchus.so
-OBJECTS = src/pam_bacchus.o src/utils.o src/tweetnacl.o
+OBJECTS = src/pam_bacchus.o src/utils.o src/tweetnacl.o src/randombytes.o
 CFLAGS = -Wall -O2 -fPIC
 LDFLAGS = -shared -Xlinker -x -lcurl
 
@@ -12,12 +12,14 @@ clean:
 	rm -f $(OBJECTS) $(TARGET)
 	rm -f pam_bacchus_genkey src/pam_bacchus_genkey.o
 
-.PHONY: test
-test: genkey
+.PHONY: testdir
+testdir:
 	$(MAKE) -C test
+
+test: $(TARGET) genkey testdir
 	./run_test.sh
 
-genkey: src/pam_bacchus_genkey.o src/tweetnacl.o
+genkey: src/pam_bacchus_genkey.o src/tweetnacl.o src/randombytes.o
 	$(CC) -o pam_bacchus_genkey $^
 
 $(TARGET): $(OBJECTS)
